refactor(backend): split index.ts into bootstrap modules (#140)
index.ts went from 179 lines of intertwined setup to 40 lines of orchestration.
Each concern now lives in a single-purpose module under src/bootstrap/:

- security.ts — CORS, cookies, JWT, rate-limit, RBAC
- docs.ts     — Swagger + UI + onRoute auto-tagging (dev-only)
- routes.ts   — 12 route groups + TMDB language cache warm
- plugins.ts  — loadAll + registerWithFastify + /api/plugins
- static.ts   — production SPA fallback (no-op in dev)
- jobs.ts     — initNotifications + startScheduler

Registration order is preserved and made explicit in index.ts: security →
docs (before routes so the onRoute hook tags everything) → routes →
initNotifications (before plugins so providers can extend) → plugins →
static → listen → scheduler.

No behavior change. Makes adding new routes / swapping auth mechanism /
extracting per-feature bootstrap work mechanical instead of diff-soup.
